fix(rust): Add file playback source files to build-source mode
Problem Analysis:
- Build Source (Dist) mode on macOS/Windows was failing with undefined symbols
- Root cause: ccap_file_reader_apple.mm and ccap_file_reader_windows.cpp were missing
- These files are required when CCAP_ENABLE_FILE_PLAYBACK is enabled (default ON in CMake)
- The FileReaderApple/Windows classes are referenced in ccap_imp_apple.mm and ccap_imp_windows.cpp

Changes:
- Add ccap_file_reader_apple.mm to macOS platform sources
- Add ccap_file_reader_windows.cpp to Windows platform sources
- Define CCAP_ENABLE_FILE_PLAYBACK=1 to enable file playback functionality

Verification:
- Linux build-source mode: ✅ cargo build --no-default-features --features build-source
- Linux tests: ✅ cargo test --no-default-features --features build-source
- All 11 tests passed

This fix ensures build-source mode includes all required source files
for the file playback feature, resolving linker errors on macOS and Windows.

+++ src/ccap_convert_frame.cpp
@@ -242,7 +242,7 @@ bool inplaceConvertFrame(VideoFrame* frame, PixelFormat toFormat, bool verticalF
     if (ret) {
         assert(frame->pixelFormat == toFormat);
         assert(frame->allocator != nullptr && frame->data[0] == frame->allocator->data());
-        frame->sizeInBytes = frame->allocator->size();
+        frame->sizeInBytes = static_cast<uint32_t>(frame->allocator->size());
     }
     return ret;
 }
